requirements.txt
python-osc
RPi.GPIO
smbus2
